requirements.txt
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11
django-bootstrap3==11.1.0
django-heroku==0.3.1
django-registration==2.4.1
django-rest-auth==0.9.5
django-rest-authtoken==1.2.2
django-tinymce==2.8.0
djangorestframework==3.10.2
gunicorn==19.9.0
Pillow==6.1.0
psycopg2==2.8.3
python-decouple==3.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
whitenoise==4.1.3